Fix type mismatch.

+++ src/config/authority.cpp
@@ -182,7 +182,7 @@ std::string authority::to_hostname() const
 message::network_address authority::to_network_address() const
 {
     static constexpr uint32_t services = 0;
-    static constexpr uint64_t timestamp = 0;
+    static constexpr uint32_t timestamp = 0;
     const message::network_address network_address
     {
         timestamp, services, ip(), port(),
